refactor(schemas,core,console): skip onboarding if user has pending invitations
