+++ packages/console/src/onboarding/hooks/use-user-onboarding-data.ts
@@ -1,4 +1,10 @@
-import { adminTenantId } from '@logto/schemas';
+import {
+  adminTenantId,
+  Project,
+  type UserOnboardingData,
+  userOnboardingDataGuard,
+  userOnboardingDataKey,
+} from '@logto/schemas';
 import { useCallback, useContext, useMemo } from 'react';
 import { z } from 'zod';
 
@@ -6,12 +12,15 @@ import { isCloud } from '@/consts/env';
 import { TenantsContext } from '@/contexts/TenantsProvider';
 import useCurrentUser from '@/hooks/use-current-user';
 
-import type { UserOnboardingData } from '../types';
-import { Project, userOnboardingDataGuard } from '../types';
-
-const userOnboardingDataKey = 'onboarding';
-
-const useUserOnboardingData = () => {
+const useUserOnboardingData = (): {
+  data: UserOnboardingData;
+  error: unknown;
+  isLoading: boolean;
+  isLoaded: boolean;
+  isOnboarding: boolean;
+  isBusinessPlan: boolean;
+  update: (data: Partial<UserOnboardingData>) => Promise<void>;
+} => {
   const { customData, error, isLoading, isLoaded, updateCustomData } = useCurrentUser();
   const { currentTenantId } = useContext(TenantsContext);
