+++ packages/schemas/src/types/index.ts
@@ -27,3 +27,4 @@ export * from './tenant-organization.js';
 export * from './mapi-proxy.js';
 export * from './consent.js';
 export * from './jwt-customizer.js';
+export * from './onboarding.js';
